Refactor: standardize structure, restore Login and Register pages, update routing
