+++ src/components/common/ErrorBoundary.jsx
@@ -1,30 +1,1 @@
-// src/components/common/ErrorBoundary.jsx
-import { Component } from "react";
 
-export default class ErrorBoundary extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { hasError: false, error: null };
-  }
-
-  static getDerivedStateFromError(error) {
-    return { hasError: true, error };
-  }
-
-  componentDidCatch(error, errorInfo) {
-    console.error("Error caught in boundary:", error, errorInfo);
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return (
-        <div className="p-8 text-center text-red-600">
-          <h2 className="text-xl font-bold">Something went wrong.</h2>
-          <p>{this.state.error?.message}</p>
-        </div>
-      );
-    }
-
-    return this.props.children;
-  }
-}
