+++ src/components/common/Loader.jsx
@@ -1,7 +1,1 @@
-export default function Loader() {
-  return (
-    <div className="flex justify-center items-center h-full w-full">
-      <div className="animate-spin rounded-full h-12 w-12 border-t-4 border-[#ffffff] border-opacity-80"></div>
-    </div>
-  );
-}
+
